重构了效果，新增了火球术测试。
