+++ Assets/TouhouHeartStone/Scripts/GameCore/Builtin/CardCategory.cs
@@ -3,7 +3,7 @@
     /// <summary>
     /// 关于ID格式，目前是8位十六进制，前3位为人物ID，中间2位是卡牌种类，后3位是卡片自己的ID。
     /// </summary>
-    class CardCategory
+    public class CardCategory
     {
         public const int MASTER = 0x00000000;
         public const int SKILL = 0x00001000;
