+++ Assets/TouhouHeartStone/Tests/TestCards.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2a6fb484e4aa25d49840de52b477aca1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
